+++ packages/cli/src/commands/package-dynamic-plugins/index.ts
@@ -1,0 +1,1 @@
+export { command } from './command';