+++ plugins/feedback/src/components/CreateFeedbackModal/index.ts
@@ -1,1 +1,0 @@
-export { CreateFeedbackModal } from './CreateFeedbackModal';