+++ plugins/feedback/src/components/EntityFeedbackPage/index.tsx
@@ -1,1 +1,0 @@
-export { EntityFeedbackPage } from './EntityFeedbackPage';