+++ plugins/feedback/src/index.ts
@@ -1,7 +1,0 @@
-export { default as FeedbackIcon } from '@mui/icons-material/TextsmsOutlined';
-export {
-  feedbackPlugin,
-  GlobalFeedbackPage,
-  EntityFeedbackPage,
-  OpcFeedbackComponent,
-} from './plugin';