+++ plugins/bulk-import-backend/CHANGELOG.md
@@ -1,5 +1,19 @@
 ### Dependencies
 
+* **@janus-idp/cli:** upgraded to 1.15.0
+
+### Dependencies
+
+* **@janus-idp/backstage-plugin-bulk-import-common:** upgraded to 1.1.0
+* **@janus-idp/cli:** upgraded to 1.14.0
+* **@janus-idp/backstage-plugin-audit-log-node:** upgraded to 1.5.0
+
+### Dependencies
+
+* **@janus-idp/cli:** upgraded to 1.13.2
+
+### Dependencies
+
 * **@janus-idp/backstage-plugin-audit-log-node:** upgraded to 1.4.1
 
 ### Dependencies