+++ plugins/sonarqube-actions/CHANGELOG.md
@@ -1,5 +1,17 @@
 ### Dependencies
 
+* **@janus-idp/cli:** upgraded to 1.15.0
+
+### Dependencies
+
+* **@janus-idp/cli:** upgraded to 1.14.0
+
+### Dependencies
+
+* **@janus-idp/cli:** upgraded to 1.13.2
+
+### Dependencies
+
 * **@janus-idp/cli:** upgraded to 1.13.1
 
 ## @janus-idp/backstage-scaffolder-backend-module-sonarqube [1.7.0](https://github.com/janus-idp/backstage-plugins/compare/@janus-idp/backstage-scaffolder-backend-module-sonarqube@1.6.0...@janus-idp/backstage-scaffolder-backend-module-sonarqube@1.7.0) (2024-07-26)